# server/ecommerce_backend/swagger.yml
openapi: 3.0.3
info:
  title: E-Commerce API
  description: |
    REST API for E-Commerce platform with user authentication, product management, 
    cart operations, and order processing.
  version: 1.0.0
  contact:
    name: API Support
    email: support@yourdomain.com
servers:
  - url: http://localhost:8000/api
    description: Development server
  - url: https://api.yourdomain.com/api
    description: Production server
...
components:
  securitySchemes:
    TokenAuth:
      type: apiKey
      in: header
      name: Authorization
      description: "Use format: Token <your_token_here>"
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: "Gaming Laptop - ROG Strix G15"
        slug:
          type: string
          example: "gaming-laptop-rog-strix-g15"
        description:
          type: string
          example: "High-performance gaming laptop with RTX 4060"
        short_description:
          type: string
          example: "Latest gen gaming laptop"
        price:
          type: string
          format: decimal
          example: "89999.00"
        old_price:
          type: string
          format: decimal
          example: "99999.00"
        discount_percentage:
          type: number
          format: float
          example: 10.0
        stock_quantity:
          type: integer
          example: 50
        main_image:
          type: string
          format: uri
          example: "http://localhost:8000/media/products/laptop.jpg"
        average_rating:
          type: number
          format: float
          example: 4.5
        review_count:
          type: integer
          example: 120
...
paths:
...
  /products/{slug}/:
    get:
      tags:
        - Products
      summary: Get product details
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          example: "gaming-laptop-rog-strix-g15"
      responses:
        '200':
          description: Product details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
...
  /admin-api/products/{id}/:
    get:
      tags:
        - Admin - Products
      summary: Get product details (admin)
      security:
        - TokenAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Product details retrieved

    put:
      tags:
        - Admin - Products
      summary: Update product
      security:
        - TokenAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: Product updated

    delete:
      tags:
        - Admin - Products
      summary: Delete product
      security:
        - TokenAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Product deleted
...